requirements.txt
Django>=3.2,<4.0
gunicorn
djangorestframework
mysqlclient
djangorestframework-simplejwt
